recommendation-api/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
scikit-learn==1.5.2
numpy==2.1.3
requests==2.32.3